Rename ContextManager to ViewContextManager
